and then, of course, only if it exists
git-svn-id: http://svn.osafoundation.org/chandler/trunk@3767 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/chandlerdb/setup.py
@@ -34,7 +34,7 @@ def main():
     setup(name='chandlerdb', version='0.4',
           ext_modules=extensions,
           py_modules=modules)
-    if os.name == 'nt':
+    if os.name == 'nt' and os.path.exists('rijndael.py'):
         shutil.move('rijndael.py', 'chandlerdb/util/rijndael.py')
     setup(name='chandlerdb', version='0.4',
           py_modules=['chandlerdb.util.rijndael'])
